+++ src/ladb_opencutlist/ruby/model/partdef.rb
@@ -4,7 +4,7 @@ module Ladb::OpenCutList
 
   class PartDef
 
-    attr_accessor :definition_id, :number, :saved_number, :name, :count, :scale, :raw_size, :size, :material_name, :material_origins, :cumulable, :orientation_locked_on_axis, :labels, :auto_oriented, :aligned_on_axes
+    attr_accessor :definition_id, :number, :saved_number, :name, :count, :scale, :raw_size, :size, :material_name, :material_origins, :cumulable, :orientation_locked_on_axis, :labels, :auto_oriented, :aligned_on_axes, :real_area
     attr_reader :id, :entity_ids, :entity_serialized_paths, :entity_names, :contains_blank_entity_names
 
     def initialize(id)
@@ -28,6 +28,7 @@ module Ladb::OpenCutList
       @contains_blank_entity_names = false
       @auto_oriented = false
       @aligned_on_axes = false
+      @real_area = nil
     end
 
     # -----